remove config functions into utils hyperstructure
